calculateBalanced, e makeTree

+++ Node.cs
@@ -1,8 +1,9 @@
+using System;
     class Node
     {
-        public int Id { get; set; }
+        public Int64 Id { get; set; }
 
-        public int RValue { get; set; }
+        public Int64 RValue { get; set; }
 
-        public int LValue { get; set; }
+        public Int64 LValue { get; set; }
     }
